Change routing and fix files names
