+++ src/App.js
@@ -1,19 +1,19 @@
 import React from 'react';
 import './App.css';
-import {Navigate, Route, Routes} from "react-router-dom";
+import {Route, Routes} from "react-router-dom";
 import Documents from "./pages/Documents/documents";
 import Assignments from "./pages/Assignments/assignments";
 import Posts from "./pages/Posts/posts";
 import SignIn from "./pages/SignIn/signIn";
 import SignUp from "./pages/SignUp/signUp";
 import Home from "./pages/Home/home";
-import News from "./pages/News/News";
-import Other from "./pages/Other/Other";
+import News from "./pages/News/news";
+import Other from "./pages/Other/other";
 
 function App() {
   return (
       <Routes>
-          <Route path="/" element={<Navigate to="/signin" replace />}/>
+          <Route path="/" element={<SignIn />}/>
           <Route path="/signIn" element={<SignIn />} />
           <Route path="/signUp" element={<SignUp />} />
           <Route path="/home" element={<Home />} />
