+++ src/components/Icon/newsIcon.jsx
