--- src/components/header/Header.css
@@ -1,75 +1,0 @@
-.header {
-  width: 100%;
-  height: 45px;
-  background-color: #001e2d;
-  position: sticky;
-  top: 0;
-  z-index: 999;
-}
-
-.topbarWrapper {
-  height: 100%;
-  padding: 0px 20px;
-  justify-content: space-between;
-  display: flex;
-  align-items: center;
-}
-.logoInfor {
-  position: relative;
-  /* background-color: blue; */
-  left: -10px;
-  justify-content: flex-start;
-  height: auto;
-  width: auto;
-}
-.logoImg {
-  border-radius: 50%;
-  position: relative;
-  top: -1px;
-}
-
-.logoText {
-  font-weight: bold;
-  font-size: 25px;
-  color: white;
-  cursor: pointer;
-  position: relative;
-  padding-left: 10px;
-  top: 4px;
-}
-
-.userOption {
-  height: 50%;
-  width: auto;
-  position: relative;
-  top: 7px;
-  padding-right: 2px;
-}
-.userImg {
-  border-radius: 50%;
-  position: relative;
-  top: 0px;
-}
-
-.userName {
-  font-weight: bold;
-  font-size: 17px;
-  color: white;
-  position: relative;
-  top: 7px;
-  padding-right: 4px;
-}
-.userInfor {
-  position: relative;
-  display: flex;
-  top: 00%;
-  justify-content: flex-end;
-  /* background-color: red; */
-  height: auto;
-  width: auto;
-}
-.children {
-  position: fixed;
-  width: 100%;
-  height: 100%;
-}
